refactor: snapshots, a masterclass, and 100% coverage
