+++ packages/slice/__tests__/android/leadandtwo.android.test.js
@@ -1,0 +1,5 @@
+import leadandtwo from "../leadandtwo";
+
+describe("LeadAndTwoSlice tests on android", () => {
+  leadandtwo();
+});
