+++ packages/slice/__tests__/android/opinionandtwo.android.test.js
@@ -1,0 +1,5 @@
+import opinionandtwo from "../opinionandtwo";
+
+describe("OpinionAndTwoSlice tests on android", () => {
+  opinionandtwo();
+});
