--- packages/slice/__tests__/android/slice.android.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("Slice tests on android", () => {
-  shared();
-});
