+++ packages/slice/__tests__/android/standard.android.test.js
@@ -1,0 +1,5 @@
+import standard from "../standard";
+
+describe("StandardSlice tests on android", () => {
+  standard();
+});
