+++ packages/slice/__tests__/ios/leadandtwo.ios.test.js
@@ -1,0 +1,5 @@
+import leadandtwo from "../leadandtwo";
+
+describe("LeadAndTwoSlice tests on ios", () => {
+  leadandtwo();
+});
