+++ packages/slice/__tests__/ios/opinionandtwo.ios.test.js
@@ -1,0 +1,5 @@
+import opinionandtwo from "../opinionandtwo";
+
+describe("OpinionAndTwoSlice tests on ios", () => {
+  opinionandtwo();
+});
