--- packages/slice/__tests__/ios/slice.ios.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("Slice tests on ios", () => {
-  shared();
-});
