+++ packages/slice/__tests__/ios/standard.ios.test.js
@@ -1,0 +1,5 @@
+import standard from "../standard";
+
+describe("StandardSlice tests on ios", () => {
+  standard();
+});
