+++ packages/slice/__tests__/web/leadandtwo.web.test.js
@@ -1,0 +1,5 @@
+import leadandtwo from "../leadandtwo";
+
+describe("LeadAndTwoSlice tests on web", () => {
+  leadandtwo();
+});
