+++ packages/slice/__tests__/web/opinionandtwo.web.test.js
@@ -1,0 +1,5 @@
+import opinionandtwo from "../opinionandtwo";
+
+describe("OpinionAndTwoSlice tests on web", () => {
+  opinionandtwo();
+});
