--- packages/slice/__tests__/web/slice.web.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("Slice tests on web", () => {
-  shared();
-});
