+++ packages/slice/__tests__/web/standard.web.test.js
@@ -1,0 +1,5 @@
+import standard from "../standard";
+
+describe("StandardSlice tests on web", () => {
+  standard();
+});
